requirements.txt
Flask==2.0.2
google-cloud-logging==2.7.0
google-cloud-firestore==2.3.4
